+++ langformers/__init__.py
@@ -6,6 +6,7 @@ from .generators import OllamaGenerator, HuggingFaceGenerator
 from .labellers import OllamaDataLabeller, HuggingFaceDataLabeller
 from .mimickers import EmbeddingMimicker
 from .searchers import FaissSearcher, ChromaDBSearcher, PineconeSearcher
+from .chunkers import FixedSizeChunker, SemanticChunker
 
 __all__ = [
     "tasks",
@@ -21,6 +22,8 @@ __all__ = [
     "FaissSearcher",
     "ChromaDBSearcher",
     "PineconeSearcher",
+    "FixedSizeChunker",
+    "SemanticChunker",
 ]