+++ docs/source/changelog.rst
@@ -6,6 +6,16 @@ All notable changes to this project are documented here.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+v0.3.0 (2024-04-14)
+---------------------
+
+Added
+^^^^^^
+- Chunking pipelines
+    - Fixed-size chunking
+    - Semantic chunking
+- Documentation and README.md updated.
+
 v0.2.0 (2024-04-10)
 ---------------------